added the validations

+++ user/src/main/java/com/deloitte/user/service/UserServiceImpl.java
@@ -2,13 +2,15 @@ package com.deloitte.user.service;
 
 import com.deloitte.user.dao.UserDao;
 import com.deloitte.user.entity.Users;
+import com.deloitte.user.exception.DuplicateResourceException;
+import com.deloitte.user.exception.InsufficientDataException;
+import com.deloitte.user.exception.RequestValidationException;
+import com.deloitte.user.exception.ResourceNotFoundException;
 import com.deloitte.user.model.UserRegistry;
 import com.deloitte.user.model.UserResponse;
 import com.deloitte.user.model.UserUpdate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 public class UserServiceImpl implements UserService{
@@ -18,35 +20,49 @@ public class UserServiceImpl implements UserService{
     @Override
     public String userRegistration(UserRegistry userRegistry) {
         //all values are intact or not
+        if (userRegistry.getFirstName()==null || userRegistry.getLastName()==null
+            || userRegistry.getEmail()==null || userRegistry.getMobileNumber()==null
+            || userRegistry.getPassword()==null || userRegistry.getConfirmationPassword()==null)
+            throw new InsufficientDataException("please enter the all the required details");
         //verify the password & conPassword
+        if (!userRegistry.getConfirmationPassword().equals(userRegistry.getPassword()))
+            throw new RequestValidationException("Passwords are not matched , please try again ");
         // email format
         //mobile number
+
         //existing email or not
 
+        if(userDao.existsUsersByEmail(userRegistry.getEmail())){
+            throw new DuplicateResourceException("user with email already taken");
+        }
         return userDao.userRegistration(userRegistry);
     }
 
     @Override
-    public UserResponse getUserById(Integer userId) {
+    public Users getUserById(Integer userId) {
 
-        Optional<Users> userById = userDao.getUserById(userId);
+       return userDao.getUserById(userId).orElseThrow(
+                () -> new ResourceNotFoundException("user with  id %s not found".formatted(userId)));
 
-        UserResponse userResponse=new UserResponse();
+//        UserResponse userResponse=new UserResponse();
 
-        if (userById!=null && !userById.isEmpty() && userById.isPresent()){
-            userResponse.setFirstName(userById.get().getFirst_name());
-            userResponse.setLastName(userById.get().getLast_name());
-            userResponse.setEmail(userById.get().getEmail());
-            userResponse.setMobileNumber(userById.get().getMobileNumber());
-        }
-        else throw new RuntimeException("no data found for the Id");
-
-        return userDao.getUserById(userId).orElseThrow(() - new RuntimeException("no data"));
+//        if (userById!=null ){
+//            userResponse.setFirstName(userById.getFirst_name());
+//            userResponse.setLastName(userById.getLast_name());
+//            userResponse.setEmail(userById.getEmail());
+//            userResponse.setMobileNumber(userById.getMobileNumber());
+//        }
+//        else throw new RuntimeException("no data found for the Id");
+//
+//        return userDao.getUserById(userId).orElseThrow();
 
     }
 
     @Override
     public String deleteUser(Integer id) {
+        if (!userDao.existsUsersByUserId(id)) {
+            throw new ResourceNotFoundException("user with id %s is not found".formatted(id));
+        }
         return userDao.deleteUser(id);
     }
 
@@ -55,16 +71,16 @@ public class UserServiceImpl implements UserService{
     @Override
     public String updateUserDetails(Integer id, UserUpdate userUpdate) {
        if ( !userDao.existsUsersByUserId(id)) return "User not found with given ID";
-        Optional<Users> users=userDao.getUserById(id);
+        Users users=userDao.getUserById(id).orElseThrow();
 
         if (userUpdate.getFirstName()!=null)
-            users.get().setFirst_name(userUpdate.getFirstName());
+            users.setFirst_name(userUpdate.getFirstName());
         if (userUpdate.getLastName()!=null)
-            users.get().setLast_name(userUpdate.getLastName());
+            users.setLast_name(userUpdate.getLastName());
         if (userUpdate.getMobileNumber()!=null)
-            users.get().setMobileNumber(userUpdate.getMobileNumber());
+            users.setMobileNumber(userUpdate.getMobileNumber());
         if (userUpdate.getEmail()!=null)
-            users.get().setEmail(userUpdate.getEmail());
+            users.setEmail(userUpdate.getEmail());
         //
         return userDao.updateUserDetails(id,users);
     }
